Move a test to elastic.rs
