+++ src/slurm.rs
@@ -69,10 +69,7 @@ pub fn is_job_path(path: &Path) -> Option<(&str, &str)> {
 mod tests {
 
     use super::*;
-    use std::env;
     use std::fs::create_dir;
-    use std::fs::File;
-    use std::io::Read;
     use tempfile::tempdir;
 
     #[test]
@@ -88,16 +85,5 @@ mod tests {
         let fdir = tdir.path().join("fubar");
         let _faildir = create_dir(&fdir);
         assert_eq!(is_job_path(&fdir), None);
-    }
-
-    #[test]
-    fn test_read_env() {
-        let path = PathBuf::from("tests/job.123456");
-        let slurm_job_entry = SlurmJobEntry::new(&path, "123456");
-        let hm = slurm_job_entry.read_env();
-
-        assert_eq!(hm.len(), 46);
-        assert_eq!(hm.get("SLURM_CLUSTERS").unwrap(), "cluster");
-        assert_eq!(hm.get("SLURM_NTASKS_PER_NODE").unwrap(), "1");
     }
 }
